update tipe data

+++ database/migrations/2020_12_22_054923_create_keuangans_table.php
@@ -18,7 +18,7 @@ class CreateKeuangansTable extends Migration
             $table->string('keterangan');
             $table->integer('debet');
             $table->integer('kredit');
-            $table->integer('saldo');
+            $table->bigInteger('saldo');
             $table->timestamps();
         });
     }
